Very far ahead planning
JW DreamCubes were a marvel of technology and holy spirit.  Several hundred years after Armageddon the devices had appeared dissembled with clear instructions at distribution centers.  The were high technology but appeared relatively simple. A wooden box with a small set of crystals that acted as tuners.  There was a slot for something similar to a cd, and the rest of the box had rather 20th century technology in it.  A small headset for the user was wireless. It ran on something called JadeOS named after the Jade and Neeta series. They had the original actress record some instructions on the basic used of the device, and she sometimes appeared as a hologram while in lucid dreams. 

It was explained in meetings that the world that hand not seemed to have made it through Amageddon were in a buffer of being ‘cut off’ neither quite dead or alive. It was hard to understand in a linear sense, but there had been deep studies into time that suggested something like this made a lot of sense.

Through the Dreamcubes they could meet people that were hard to reach. Those in warzones, Hollywood, scientists and others who scoffed at the Witnesses, were unable to be reached or needed something more scientific to talk about to make common ground.  Witnesses could select the type of person they wanted to find, how they would meet them, and what supplies they could have with them.

There was some talk of an Entertainment Division who used a different OS called Mini-Chariot.  One that referenced archtypes and something called Narrative physics.  There was another OS called Enigma 2 that researched real time travel and connected to something everyone just called Computer Zero.  New documents appeared at random times, piles of code and other artifacts worth looking at. Some quick covers for a Watchtower and Awake had been found and ‘moved’ into the future. It was a good time to have them as they filled a small gap where everything was being retooled.